+++ custom_components/warema_wms/__init__.py
@@ -27,6 +27,7 @@ PLATFORMS: list[Platform] = [
     Platform.BINARY_SENSOR,
     Platform.BUTTON,
     Platform.COVER,
+    Platform.LIGHT,
     Platform.SENSOR,
 ]